style(gedzip): drop unnecessary int64() conversion in test
maxGEDZIPEntryBytes is already int64 (see glx/gedzip_import.go:46),
so int64(maxGEDZIPEntryBytes) is a same-type cast that trips the
unconvert linter. Drop the wrapper.

+++ glx/cmd_import_gedzip_test.go
@@ -466,7 +466,7 @@ func TestWriteZipEntry_AcceptsEntryAtExactLimit(t *testing.T) {
 
 	info, statErr := os.Stat(destPath)
 	require.NoError(t, statErr, "extracted file must exist")
-	require.Equal(t, int64(maxGEDZIPEntryBytes), info.Size())
+	require.Equal(t, maxGEDZIPEntryBytes, info.Size())
 }
 
 func TestEntrySizeLimitReader(t *testing.T) {
